@apelisse 2019-09-25
Integrating this change in Kubernetes:
```
benchmark                                           old ns/op     new ns/op     delta
BenchmarkApplyNewObject//v1,_Kind=Pod-12            2292629       2271609       -0.92%
BenchmarkApplyNewObject//v1,_Kind=Node-12           4676879       4462379       -4.59%
BenchmarkApplyNewObject//v1,_Kind=Endpoints-12      89657544      88854655      -0.90%
BenchmarkUpdateNewObject//v1,_Kind=Pod-12           1115609       771370        -30.86%
BenchmarkUpdateNewObject//v1,_Kind=Node-12          1584057       1086912       -31.38%
BenchmarkUpdateNewObject//v1,_Kind=Endpoints-12     6047963       4073366       -32.65%
BenchmarkRepeatedUpdate-12                          959051        633715        -33.92%

benchmark                                           old allocs     new allocs     delta
BenchmarkApplyNewObject//v1,_Kind=Pod-12            8819           8325           -5.60%
BenchmarkApplyNewObject//v1,_Kind=Node-12           17696          16867          -4.68%
BenchmarkApplyNewObject//v1,_Kind=Endpoints-12      409590         401522         -1.97%
BenchmarkUpdateNewObject//v1,_Kind=Pod-12           3977           2982           -25.02%
BenchmarkUpdateNewObject//v1,_Kind=Node-12          5742           4290           -25.29%
BenchmarkUpdateNewObject//v1,_Kind=Endpoints-12     21441          13339          -37.79%
BenchmarkRepeatedUpdate-12                          3027           2026           -33.07%
```