@apelisse 2019-09-26
I think I've addressed the comments, still the same imrpovements overall.

For Get, having the hashmap gets better for maps bigger than ~100 items, which are I think rare. We could do something better if needed for these case but I"m still hoping to find some other improvements.